+++ cmake/ccf_app.cmake
@@ -22,12 +22,7 @@ message(STATUS "Compile target platform: ${COMPILE_TARGET}")
 include(${CCF_DIR}/cmake/open_enclave.cmake)
 
 list(APPEND COMPILE_LIBCXX -stdlib=libc++)
-if(CMAKE_CXX_COMPILER_VERSION VERSION_GREATER 9)
-  list(APPEND LINK_LIBCXX -lc++ -lc++abi -stdlib=libc++)
-else()
-  # Clang <9 needs to link libc++fs when using <filesystem>
-  list(APPEND LINK_LIBCXX -lc++ -lc++abi -lc++fs -stdlib=libc++)
-endif()
+list(APPEND LINK_LIBCXX -lc++ -lc++abi -stdlib=libc++)
 
 # Sign a built enclave library with oesign
 function(sign_app_library name app_oe_conf_path enclave_sign_key_path)