+++ tests/e2e_operations.py
@@ -318,8 +318,10 @@ def run_configuration_file_checks(args):
 
     for config in config_files_to_check:
         cmd = [bin_path, f"--config={config}", "--check"]
-        rc = infra.proc.ccall(*cmd).returncode
-        assert rc == 0, f"Failed to run tutorial script: {rc}"
+        rc = infra.proc.ccall(
+            *cmd, env={"ASAN_OPTIONS": "alloc_dealloc_mismatch=0"}
+        ).returncode
+        assert rc == 0, f"Failed to check configuration: {rc}"
 
 
 def run(args):