+++ tests/expired_service/common/actions.js
@@ -8,7 +8,7 @@ class Action {
 function parseUrl(url) {
   // From https://tools.ietf.org/html/rfc3986#appendix-B
   const re = new RegExp(
-    "^(([^:/?#]+):)?(//([^/?#]*))?([^?#]*)(\\?([^#]*))?(#(.*))?"
+    "^(([^:/?#]+):)?(//([^/?#]*))?([^?#]*)(\\?([^#]*))?(#(.*))?",
   );
   const groups = url.match(re);
   if (!groups) {
@@ -132,7 +132,7 @@ function checkJwks(value, field) {
 function checkX509CertBundle(value, field) {
   if (!ccf.isValidX509CertBundle(value)) {
     throw new Error(
-      `${field} must be a valid X509 certificate (bundle) in PEM format`
+      `${field} must be a valid X509 certificate (bundle) in PEM format`,
     );
   }
 }
@@ -153,7 +153,7 @@ function invalidateOtherOpenProposals(proposalIdToRetain) {
 
 function setServiceCertificateValidityPeriod(validFrom, validityPeriodDays) {
   const rawConfig = ccf.kv["public:ccf.gov.service.config"].get(
-    getSingletonKvKey()
+    getSingletonKvKey(),
   );
   if (rawConfig === undefined) {
     throw new Error("Service configuration could not be found");
@@ -170,13 +170,13 @@ function setServiceCertificateValidityPeriod(validFrom, validityPeriodDays) {
     validityPeriodDays > max_allowed_cert_validity_period_days
   ) {
     throw new Error(
-      `Validity period ${validityPeriodDays} (days) is not allowed: service max allowed is ${max_allowed_cert_validity_period_days} (days)`
+      `Validity period ${validityPeriodDays} (days) is not allowed: service max allowed is ${max_allowed_cert_validity_period_days} (days)`,
     );
   }
 
   const renewed_service_certificate = ccf.network.generateNetworkCertificate(
     validFrom,
-    validityPeriodDays ?? max_allowed_cert_validity_period_days
+    validityPeriodDays ?? max_allowed_cert_validity_period_days,
   );
 
   const serviceInfoTable = "public:ccf.gov.service.info";
@@ -189,7 +189,7 @@ function setServiceCertificateValidityPeriod(validFrom, validityPeriodDays) {
   serviceInfo.cert = renewed_service_certificate;
   ccf.kv[serviceInfoTable].set(
     getSingletonKvKey(),
-    ccf.jsonCompatibleToBuf(serviceInfo)
+    ccf.jsonCompatibleToBuf(serviceInfo),
   );
 }
 
@@ -197,14 +197,14 @@ function setNodeCertificateValidityPeriod(
   nodeId,
   nodeInfo,
   validFrom,
-  validityPeriodDays
+  validityPeriodDays,
 ) {
   if (nodeInfo.certificate_signing_request === undefined) {
     throw new Error(`Node ${nodeId} has no certificate signing request`);
   }
 
   const rawConfig = ccf.kv["public:ccf.gov.service.config"].get(
-    getSingletonKvKey()
+    getSingletonKvKey(),
   );
   if (rawConfig === undefined) {
     throw new Error("Service configuration could not be found");
@@ -221,18 +221,18 @@ function setNodeCertificateValidityPeriod(
     validityPeriodDays > max_allowed_cert_validity_period_days
   ) {
     throw new Error(
-      `Validity period ${validityPeriodDays} (days) is not allowed: service max allowed is ${max_allowed_cert_validity_period_days} (days)`
+      `Validity period ${validityPeriodDays} (days) is not allowed: service max allowed is ${max_allowed_cert_validity_period_days} (days)`,
     );
   }
 
   const endorsed_node_cert = ccf.network.generateEndorsedCertificate(
     nodeInfo.certificate_signing_request,
     validFrom,
-    validityPeriodDays ?? max_allowed_cert_validity_period_days
+    validityPeriodDays ?? max_allowed_cert_validity_period_days,
   );
   ccf.kv["public:ccf.gov.nodes.endorsed_certificates"].set(
     ccf.strToBuf(nodeId),
-    ccf.strToBuf(endorsed_node_cert)
+    ccf.strToBuf(endorsed_node_cert),
   );
 }
 
@@ -253,13 +253,13 @@ function checkRecoveryThreshold(config, new_config) {
 
   if (service.status === "WaitingForRecoveryShares") {
     throw new Error(
-      `Cannot set recovery threshold if service is ${service.status}`
+      `Cannot set recovery threshold if service is ${service.status}`,
     );
   } else if (service.status === "Open") {
     let activeRecoveryMembersCount = getActiveRecoveryMembersCount();
     if (new_config.recovery_threshold > activeRecoveryMembersCount) {
       throw new Error(
-        `Cannot set recovery threshold to ${new_config.recovery_threshold}: recovery threshold would be greater than the number of recovery members ${activeRecoveryMembersCount}`
+        `Cannot set recovery threshold to ${new_config.recovery_threshold}: recovery threshold would be greater than the number of recovery members ${activeRecoveryMembersCount}`,
       );
     }
   }
@@ -276,7 +276,7 @@ function checkReconfigurationType(config, new_config) {
       )
     ) {
       throw new Error(
-        `Cannot change reconfiguration type from ${from} to ${to}.`
+        `Cannot change reconfiguration type from ${from} to ${to}.`,
       );
     }
   }
@@ -310,7 +310,7 @@ function updateServiceConfig(new_config) {
 
   ccf.kv[service_config_table].set(
     getSingletonKvKey(),
-    ccf.jsonCompatibleToBuf(config)
+    ccf.jsonCompatibleToBuf(config),
   );
 
   if (need_recovery_threshold_refresh) {
@@ -328,12 +328,12 @@ const actions = new Map([
       function (args, proposalId) {
         ccf.kv["public:ccf.gov.constitution"].set(
           getSingletonKvKey(),
-          ccf.jsonCompatibleToBuf(args.constitution)
+          ccf.jsonCompatibleToBuf(args.constitution),
         );
 
         // Changing the constitution changes the semantics of any other open proposals, so invalidate them to avoid confusion or malicious vote modification
         invalidateOtherOpenProposals(proposalId);
-      }
+      },
     ),
   ],
   [
@@ -351,17 +351,17 @@ const actions = new Map([
 
         ccf.kv["public:ccf.gov.members.certs"].set(
           rawMemberId,
-          ccf.strToBuf(args.cert)
+          ccf.strToBuf(args.cert),
         );
 
         if (args.encryption_pub_key == null) {
           ccf.kv["public:ccf.gov.members.encryption_public_keys"].delete(
-            rawMemberId
+            rawMemberId,
           );
         } else {
           ccf.kv["public:ccf.gov.members.encryption_public_keys"].set(
             rawMemberId,
-            ccf.strToBuf(args.encryption_pub_key)
+            ccf.strToBuf(args.encryption_pub_key),
           );
         }
 
@@ -370,11 +370,11 @@ const actions = new Map([
         member_info.status = "Accepted";
         ccf.kv["public:ccf.gov.members.info"].set(
           rawMemberId,
-          ccf.jsonCompatibleToBuf(member_info)
+          ccf.jsonCompatibleToBuf(member_info),
         );
 
         const rawSignature = ccf.kv["public:ccf.internal.signatures"].get(
-          getSingletonKvKey()
+          getSingletonKvKey(),
         );
         if (rawSignature === undefined) {
           ccf.kv["public:ccf.gov.members.acks"].set(rawMemberId);
@@ -384,10 +384,10 @@ const actions = new Map([
           ack.state_digest = signature.root;
           ccf.kv["public:ccf.gov.members.acks"].set(
             rawMemberId,
-            ccf.jsonCompatibleToBuf(ack)
+            ccf.jsonCompatibleToBuf(ack),
           );
         }
-      }
+      },
     ),
   ],
   [
@@ -418,7 +418,7 @@ const actions = new Map([
         // to recover the service
         if (isActiveMember && isRecoveryMember) {
           const rawConfig = ccf.kv["public:ccf.gov.service.config"].get(
-            getSingletonKvKey()
+            getSingletonKvKey(),
           );
           if (rawConfig === undefined) {
             throw new Error("Service configuration could not be found");
@@ -429,14 +429,14 @@ const actions = new Map([
             getActiveRecoveryMembersCount() - 1;
           if (activeRecoveryMembersCountAfter < config.recovery_threshold) {
             throw new Error(
-              `Number of active recovery members (${activeRecoveryMembersCountAfter}) would be less than recovery threshold (${config.recovery_threshold})`
+              `Number of active recovery members (${activeRecoveryMembersCountAfter}) would be less than recovery threshold (${config.recovery_threshold})`,
             );
           }
         }
 
         ccf.kv["public:ccf.gov.members.info"].delete(rawMemberId);
         ccf.kv["public:ccf.gov.members.encryption_public_keys"].delete(
-          rawMemberId
+          rawMemberId,
         );
         ccf.kv["public:ccf.gov.members.certs"].delete(rawMemberId);
         ccf.kv["public:ccf.gov.members.acks"].delete(rawMemberId);
@@ -448,7 +448,7 @@ const actions = new Map([
           // remaining active recovery members
           ccf.node.triggerLedgerRekey();
         }
-      }
+      },
     ),
   ],
   [
@@ -469,7 +469,7 @@ const actions = new Map([
         let mi = ccf.bufToJsonCompatible(member_info);
         mi.member_data = args.member_data;
         members_info.set(member_id, ccf.jsonCompatibleToBuf(mi));
-      }
+      },
     ),
   ],
   [
@@ -485,18 +485,18 @@ const actions = new Map([
 
         ccf.kv["public:ccf.gov.users.certs"].set(
           rawUserId,
-          ccf.strToBuf(args.cert)
+          ccf.strToBuf(args.cert),
         );
 
         if (args.user_data !== null && args.user_data !== undefined) {
           ccf.kv["public:ccf.gov.users.info"].set(
             rawUserId,
-            ccf.jsonCompatibleToBuf(args.user_data)
+            ccf.jsonCompatibleToBuf(args.user_data),
           );
         } else {
           ccf.kv["public:ccf.gov.users.info"].delete(rawUserId);
         }
-      }
+      },
     ),
   ],
   [
@@ -509,7 +509,7 @@ const actions = new Map([
         const user_id = ccf.strToBuf(args.user_id);
         ccf.kv["public:ccf.gov.users.certs"].delete(user_id);
         ccf.kv["public:ccf.gov.users.info"].delete(user_id);
-      }
+      },
     ),
   ],
   [
@@ -527,12 +527,12 @@ const actions = new Map([
           userInfo.user_data = args.user_data;
           ccf.kv["public:ccf.gov.users.info"].set(
             userId,
-            ccf.jsonCompatibleToBuf(userInfo)
+            ccf.jsonCompatibleToBuf(userInfo),
           );
         } else {
           ccf.kv["public:ccf.gov.users.info"].delete(userId);
         }
-      }
+      },
     ),
   ],
   [
@@ -544,7 +544,7 @@ const actions = new Map([
       },
       function (args) {
         updateServiceConfig(args);
-      }
+      },
     ),
   ],
   [
@@ -555,7 +555,7 @@ const actions = new Map([
       },
       function (args) {
         ccf.node.triggerRecoverySharesRefresh();
-      }
+      },
     ),
   ],
   [
@@ -567,7 +567,7 @@ const actions = new Map([
 
       function (args) {
         ccf.node.triggerLedgerRekey();
-      }
+      },
     ),
   ],
   [
@@ -577,14 +577,14 @@ const actions = new Map([
         checkType(
           args.next_service_identity,
           "string",
-          "next service identity (PEM certificate)"
+          "next service identity (PEM certificate)",
         );
         checkX509CertBundle(args.next_service_identity, "cert");
 
         checkType(
           args.previous_service_identity,
           "string?",
-          "previous service identity (PEM certificate)"
+          "previous service identity (PEM certificate)",
         );
         if (args.previous_service_identity !== undefined) {
           checkX509CertBundle(args.previous_service_identity, "cert");
@@ -606,7 +606,7 @@ const actions = new Map([
             args.next_service_identity === undefined)
         ) {
           throw new Error(
-            `Opening a recovering network requires both, the previous and the next service identity`
+            `Opening a recovering network requires both, the previous and the next service identity`,
           );
         }
 
@@ -616,7 +616,7 @@ const actions = new Map([
             : undefined;
         const next_identity = ccf.strToBuf(args.next_service_identity);
         ccf.node.transitionServiceToOpen(previous_identity, next_identity);
-      }
+      },
     ),
   ],
   [
@@ -638,7 +638,7 @@ const actions = new Map([
         checkType(bundle.metadata, "object", prefix);
         checkType(bundle.metadata.endpoints, "object", `${prefix}.endpoints`);
         for (const [url, endpoint] of Object.entries(
-          bundle.metadata.endpoints
+          bundle.metadata.endpoints,
         )) {
           checkType(endpoint, "object", `${prefix}.endpoints["${url}"]`);
           for (const [method, info] of Object.entries(endpoint)) {
@@ -649,23 +649,23 @@ const actions = new Map([
             checkEnum(
               info.mode,
               ["readwrite", "readonly", "historical"],
-              `${prefix2}.mode`
+              `${prefix2}.mode`,
             );
             checkEnum(
               info.forwarding_required,
               ["sometimes", "always", "never"],
-              `${prefix2}.forwarding_required`
+              `${prefix2}.forwarding_required`,
             );
             checkType(info.openapi, "object?", `${prefix2}.openapi`);
             checkType(
               info.openapi_hidden,
               "boolean?",
-              `${prefix2}.openapi_hidden`
+              `${prefix2}.openapi_hidden`,
             );
             checkType(
               info.authn_policies,
               "array",
-              `${prefix2}.authn_policies`
+              `${prefix2}.authn_policies`,
             );
             for (const [i, policy] of info.authn_policies.entries()) {
               checkType(policy, "string", `${prefix2}.authn_policies[${i}]`);
@@ -679,7 +679,7 @@ const actions = new Map([
         checkType(
           args.disable_bytecode_cache,
           "boolean?",
-          "disable_bytecode_cache"
+          "disable_bytecode_cache",
         );
       },
       function (args) {
@@ -708,7 +708,7 @@ const actions = new Map([
         }
 
         for (const [url, endpoint] of Object.entries(
-          bundle.metadata.endpoints
+          bundle.metadata.endpoints,
         )) {
           for (const [method, info] of Object.entries(endpoint)) {
             const key = `${method.toUpperCase()} ${url}`;
@@ -719,7 +719,7 @@ const actions = new Map([
             endpointsMap.set(keyBuf, infoBuf);
           }
         }
-      }
+      },
     ),
   ],
   [
@@ -737,7 +737,7 @@ const actions = new Map([
         modulesQuickJsBytecodeMap.clear();
         modulesQuickJsVersionVal.clear();
         endpointsMap.clear();
-      }
+      },
     ),
   ],
   [
@@ -746,7 +746,7 @@ const actions = new Map([
       function (args) {},
       function (args) {
         ccf.refreshAppBytecodeCache();
-      }
+      },
     ),
   ],
   [
@@ -762,7 +762,7 @@ const actions = new Map([
         const nameBuf = ccf.strToBuf(name);
         const bundleBuf = ccf.jsonCompatibleToBuf(bundle);
         ccf.kv["public:ccf.gov.tls.ca_cert_bundles"].set(nameBuf, bundleBuf);
-      }
+      },
     ),
   ],
   [
@@ -775,7 +775,7 @@ const actions = new Map([
         const name = args.name;
         const nameBuf = ccf.strToBuf(name);
         ccf.kv["public:ccf.gov.tls.ca_cert_bundles"].delete(nameBuf);
-      }
+      },
     ),
   ],
   [
@@ -791,11 +791,11 @@ const actions = new Map([
           checkType(
             args.key_policy.sgx_claims,
             "object?",
-            "key_policy.sgx_claims"
+            "key_policy.sgx_claims",
           );
           if (args.key_policy.sgx_claims) {
             for (const [name, value] of Object.entries(
-              args.key_policy.sgx_claims
+              args.key_policy.sgx_claims,
             )) {
               checkType(value, "string", `key_policy.sgx_claims["${name}"]`);
             }
@@ -808,7 +808,7 @@ const actions = new Map([
         if (args.auto_refresh) {
           if (!args.ca_cert_bundle_name) {
             throw new Error(
-              "ca_cert_bundle_name is missing but required if auto_refresh is true"
+              "ca_cert_bundle_name is missing but required if auto_refresh is true",
             );
           }
           let url;
@@ -819,12 +819,12 @@ const actions = new Map([
           }
           if (url.scheme != "https") {
             throw new Error(
-              "issuer must be a URL starting with https:// if auto_refresh is true"
+              "issuer must be a URL starting with https:// if auto_refresh is true",
             );
           }
           if (url.query || url.fragment) {
             throw new Error(
-              "issuer must be a URL without query/fragment if auto_refresh is true"
+              "issuer must be a URL without query/fragment if auto_refresh is true",
             );
           }
         }
@@ -835,11 +835,11 @@ const actions = new Map([
           const caCertBundleNameBuf = ccf.strToBuf(args.ca_cert_bundle_name);
           if (
             !ccf.kv["public:ccf.gov.tls.ca_cert_bundles"].has(
-              caCertBundleNameBuf
+              caCertBundleNameBuf,
             )
           ) {
             throw new Error(
-              `No CA cert bundle found with name '${caCertBundleName}'`
+              `No CA cert bundle found with name '${caCertBundleName}'`,
             );
           }
         }
@@ -853,7 +853,7 @@ const actions = new Map([
         const issuerBuf = ccf.strToBuf(issuer);
         const metadataBuf = ccf.jsonCompatibleToBuf(metadata);
         ccf.kv["public:ccf.gov.jwt.issuers"].set(issuerBuf, metadataBuf);
-      }
+      },
     ),
   ],
   [
@@ -873,7 +873,7 @@ const actions = new Map([
         const metadata = ccf.bufToJsonCompatible(metadataBuf);
         const jwks = args.jwks;
         ccf.setJwtPublicSigningKeys(issuer, metadata, jwks);
-      }
+      },
     ),
   ],
   [
@@ -889,7 +889,7 @@ const actions = new Map([
         }
         ccf.kv["public:ccf.gov.jwt.issuers"].delete(issuerBuf);
         ccf.removeJwtPublicSigningKeys(args.issuer);
-      }
+      },
     ),
   ],
   [
@@ -905,7 +905,7 @@ const actions = new Map([
 
         // Adding a new allowed code ID changes the semantics of any other open proposals, so invalidate them to avoid confusion or malicious vote modification
         invalidateOtherOpenProposals(proposalId);
-      }
+      },
     ),
   ],
   [
@@ -924,7 +924,7 @@ const actions = new Map([
         let ni = ccf.bufToJsonCompatible(node_info);
         ni.node_data = args.node_data;
         nodes_info.set(node_id, ccf.jsonCompatibleToBuf(ni));
-      }
+      },
     ),
   ],
   [
@@ -937,26 +937,26 @@ const actions = new Map([
           checkType(
             args.validity_period_days,
             "integer",
-            "validity_period_days"
+            "validity_period_days",
           );
           checkBounds(
             args.validity_period_days,
             1,
             null,
-            "validity_period_days"
+            "validity_period_days",
           );
         }
       },
       function (args) {
         const rawConfig = ccf.kv["public:ccf.gov.service.config"].get(
-          getSingletonKvKey()
+          getSingletonKvKey(),
         );
         if (rawConfig === undefined) {
           throw new Error("Service configuration could not be found");
         }
         const serviceConfig = ccf.bufToJsonCompatible(rawConfig);
         const node = ccf.kv["public:ccf.gov.nodes.info"].get(
-          ccf.strToBuf(args.node_id)
+          ccf.strToBuf(args.node_id),
         );
         if (node === undefined) {
           throw new Error(`No such node: ${args.node_id}`);
@@ -971,7 +971,7 @@ const actions = new Map([
             ccf.network.getLatestLedgerSecretSeqno();
           ccf.kv["public:ccf.gov.nodes.info"].set(
             ccf.strToBuf(args.node_id),
-            ccf.jsonCompatibleToBuf(nodeInfo)
+            ccf.jsonCompatibleToBuf(nodeInfo),
           );
 
           // Also generate and record service-endorsed node certificate from node CSR
@@ -989,22 +989,23 @@ const actions = new Map([
               args.validity_period_days > max_allowed_cert_validity_period_days
             ) {
               throw new Error(
-                `Validity period ${args.validity_period_days} is not allowed: max allowed is ${max_allowed_cert_validity_period_days}`
+                `Validity period ${args.validity_period_days} is not allowed: max allowed is ${max_allowed_cert_validity_period_days}`,
               );
             }
 
             const endorsed_node_cert = ccf.network.generateEndorsedCertificate(
               nodeInfo.certificate_signing_request,
               args.valid_from,
-              args.validity_period_days ?? max_allowed_cert_validity_period_days
+              args.validity_period_days ??
+                max_allowed_cert_validity_period_days,
             );
             ccf.kv["public:ccf.gov.nodes.endorsed_certificates"].set(
               ccf.strToBuf(args.node_id),
-              ccf.strToBuf(endorsed_node_cert)
+              ccf.strToBuf(endorsed_node_cert),
             );
           }
         }
-      }
+      },
     ),
   ],
   [
@@ -1016,7 +1017,7 @@ const actions = new Map([
       function (args) {
         const codeId = ccf.strToBuf(args.code_id);
         ccf.kv["public:ccf.gov.nodes.code_ids"].delete(codeId);
-      }
+      },
     ),
   ],
   [
@@ -1027,14 +1028,14 @@ const actions = new Map([
       },
       function (args) {
         const rawConfig = ccf.kv["public:ccf.gov.service.config"].get(
-          getSingletonKvKey()
+          getSingletonKvKey(),
         );
         if (rawConfig === undefined) {
           throw new Error("Service configuration could not be found");
         }
         const serviceConfig = ccf.bufToJsonCompatible(rawConfig);
         const node = ccf.kv["public:ccf.gov.nodes.info"].get(
-          ccf.strToBuf(args.node_id)
+          ccf.strToBuf(args.node_id),
         );
         if (node === undefined) {
           return;
@@ -1042,7 +1043,7 @@ const actions = new Map([
         const node_obj = ccf.bufToJsonCompatible(node);
         if (node_obj.status === "Pending") {
           ccf.kv["public:ccf.gov.nodes.info"].delete(
-            ccf.strToBuf(args.node_id)
+            ccf.strToBuf(args.node_id),
           );
         } else {
           node_obj.status =
@@ -1051,10 +1052,10 @@ const actions = new Map([
               : "Retired";
           ccf.kv["public:ccf.gov.nodes.info"].set(
             ccf.strToBuf(args.node_id),
-            ccf.jsonCompatibleToBuf(node_obj)
+            ccf.jsonCompatibleToBuf(node_obj),
           );
         }
-      }
+      },
     ),
   ],
   [
@@ -1067,19 +1068,19 @@ const actions = new Map([
           checkType(
             args.validity_period_days,
             "integer",
-            "validity_period_days"
+            "validity_period_days",
           );
           checkBounds(
             args.validity_period_days,
             1,
             null,
-            "validity_period_days"
+            "validity_period_days",
           );
         }
       },
       function (args) {
         const node = ccf.kv["public:ccf.gov.nodes.info"].get(
-          ccf.strToBuf(args.node_id)
+          ccf.strToBuf(args.node_id),
         );
         if (node === undefined) {
           throw new Error(`No such node: ${args.node_id}`);
@@ -1093,9 +1094,9 @@ const actions = new Map([
           args.node_id,
           nodeInfo,
           args.valid_from,
-          args.validity_period_days
+          args.validity_period_days,
         );
-      }
+      },
     ),
   ],
   [
@@ -1107,13 +1108,13 @@ const actions = new Map([
           checkType(
             args.validity_period_days,
             "integer",
-            "validity_period_days"
+            "validity_period_days",
           );
           checkBounds(
             args.validity_period_days,
             1,
             null,
-            "validity_period_days"
+            "validity_period_days",
           );
         }
       },
@@ -1126,11 +1127,11 @@ const actions = new Map([
               nodeId,
               nodeInfo,
               args.valid_from,
-              args.validity_period_days
+              args.validity_period_days,
             );
           }
         });
-      }
+      },
     ),
   ],
   [
@@ -1142,22 +1143,22 @@ const actions = new Map([
           checkType(
             args.validity_period_days,
             "integer",
-            "validity_period_days"
+            "validity_period_days",
           );
           checkBounds(
             args.validity_period_days,
             1,
             null,
-            "validity_period_days"
+            "validity_period_days",
           );
         }
       },
       function (args) {
         setServiceCertificateValidityPeriod(
           args.valid_from,
-          args.validity_period_days
+          args.validity_period_days,
         );
-      }
+      },
     ),
   ],
   [
@@ -1167,7 +1168,7 @@ const actions = new Map([
         for (var key in args) {
           if (key !== "reconfiguration_type" && key !== "recovery_threshold") {
             throw new Error(
-              `Cannot change ${key} via set_service_configuration.`
+              `Cannot change ${key} via set_service_configuration.`,
             );
           }
         }
@@ -1177,7 +1178,7 @@ const actions = new Map([
       },
       function (args) {
         updateServiceConfig(args);
-      }
+      },
     ),
   ],
   [
@@ -1186,7 +1187,7 @@ const actions = new Map([
       function (args) {},
       function (args, proposalId) {
         ccf.node.triggerLedgerChunk();
-      }
+      },
     ),
   ],
   [
@@ -1195,7 +1196,7 @@ const actions = new Map([
       function (args) {},
       function (args, proposalId) {
         ccf.node.triggerSnapshot();
-      }
+      },
     ),
   ],
 ]);