+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.13]
+
+[3.0.13]: https://github.com/microsoft/CCF/releases/tag/ccf-3.0.13
+
+- Updated Open Enclave to [0.19.3](https://github.com/openenclave/openenclave/releases/tag/v0.19.3) (#5472).
+- Updated to `clang-11` for SGX builds (#5472).
+- Updated to `clang-15` for Virtual builds (#5472).
+
 ## [3.0.12]
 
 [3.0.12]: https://github.com/microsoft/CCF/releases/tag/ccf-3.0.12